close #49: Standardize the place of the line in error and warning messages
