+++ assets/TOOLS.md
@@ -9,8 +9,8 @@ Mensagem:
 close #numero_da_issue: descrição do commit
 
 # Exemplo:
-Issue: Tamanho do titulo #35 Emitir warning para todo nome e titulo que tiver mais do que 30 caracteres. Adicionar flag para desabilitar este tipo de warning
+Issue: Padronizar lugar da linha nas mensagens de erro e warning #49
 
 
 Escrever o nome do commit em inglês e usar o modo imperativo.
-close #35: Add flag to disable warning for long titles and names
+close #49: Standardize the place of the line in error and warning messages
