+++ tests/unit/test_constants.py
@@ -1,0 +1,3 @@
+# Diretórios de entrada para os testes
+path_input_data_ground_truth = "input_data/data_ground_truth"
+path_input_data_errors = "input_data/data_errors"
