--- tests/unit/test_count_errors.py
@@ -1,98 +1,0 @@
-from src.orchestrator import verify_structure_folder_files
-from src.orchestrator import verify_sp_description_parser
-from src.orchestrator import verify_sp_description_titles_uniques
-from src.orchestrator import verify_sp_description_text_capitalize
-from src.orchestrator import verify_graph_sp_description_composition
-from src.orchestrator import verify_ids_sp_description_values
-from src.myparser.sp_description import verify_sp_description_titles_length
-from src.myparser.spellchecker import TypeDict
-from src.orchestrator import verify_spelling_text
-
-
-# Diretórios de entrada para os testes
-path_input_data_ground_truth = "input_data/data_ground_truth"
-path_input_data_errors = "input_data/data_errors"
-
-def test_count_errors_verify_sp_description_titles_length(): # Teste false
-    planilha_04_descricao = path_input_data_errors + "/4_descricao/descricao.xlsx"
-    is_correct, errors, warnings = verify_sp_description_titles_length(planilha_04_descricao)
-    assert is_correct is True
-    # Numero de erros esperado == 0
-    assert len(errors) == 0
-    # Numero de warnings esperado == 1
-    assert len(warnings) == 1
-
-
-# Testes: Issue #5: Códigos html nas descrições simples
-def test_count_errors_verify_sp_description_parser(): # Teste false
-    planilha_04_descricao = path_input_data_errors + "/4_descricao/descricao.xlsx"
-    is_correct, errors, warnings = verify_sp_description_parser(planilha_04_descricao)
-    # Numero de erros esperado == 3
-    assert len(errors) == 3
-    # Numero de warnings esperado == 0
-    assert len(warnings) == 0
-
-# Testes: Issue #39: Estrutura da pasta de arquivos
-def test_count_errors_verify_structure_folder_files(): # Teste false
-    is_correct, errors, warnings = verify_structure_folder_files(path_input_data_errors)
-    # Numero de erros esperado == 2
-    assert len(errors) == 2
-    # Numero de warnings esperado == 0
-    assert len(warnings) == 0
-    
-# Testes: Issue #36: Títulos únicos
-def test_count_errors_verify_sp_description_titles_uniques(): # Teste false
-    planilha_04_descricao = path_input_data_errors + "/4_descricao/descricao.xlsx"
-    is_correct, errors, warnings = verify_sp_description_titles_uniques(planilha_04_descricao)
-    # Numero de erros esperado == 2
-    assert len(errors) == 2
-    # Numero de warnings esperado == 0
-    assert len(warnings) == 0
-
-# Testes: Issue #1: Padrão para nomes dos indicadores
-def test_count_errors_verify_sp_description_text_capitalize(): # Teste false
-    planilha_04_descricao = path_input_data_errors + "/4_descricao/descricao.xlsx"
-    is_correct, errors, warnings = verify_sp_description_text_capitalize(planilha_04_descricao)
-    # Numero de warnings esperado == 10
-    assert len(warnings) == 10
-    # Numero de erros esperado == 0
-    assert len(errors) == 0
-
-# Testes: 6 - Hierarquia como grafo conexo
-def test_count_errors_verify_graph_sp_description_composition(): # Teste false
-    planilha_04_descricao = path_input_data_errors + "/4_descricao/descricao.xlsx"
-    planilha_05_composicao = path_input_data_errors + "/5_composicao/composicao.xlsx"
-    is_correct, errors, warnings = verify_graph_sp_description_composition(planilha_04_descricao, planilha_05_composicao)
-    # Numero de erros esperado == 4
-    assert len(errors) == 4
-    # Numero de warnings esperado == 0
-    assert len(warnings) == 0
-
-# Testes: 2.2 - Verificar indicadores dos arquivos descricao.xlsx ausentes em valores.xlsx, e vice-versa
-def test_count_errors_verify_ids_sp_description_values(): # Teste false
-    planilha_04_descricao = path_input_data_errors + "/4_descricao/descricao.xlsx"
-    planilha_08_valores = path_input_data_errors + "/8_valores/valores.xlsx"
-    is_correct, errors, warnings = verify_ids_sp_description_values(planilha_04_descricao, planilha_08_valores)
-    # Numero de erros esperado == 2
-    assert len(errors) == 2
-    # Numero de warnings esperado == 0
-    assert len(warnings) == 0
-
-def test_count_errors_verify_spelling_text_tiny(): # Teste false
-    path_input_folder = path_input_data_errors
-    type_dict_spell = TypeDict.TINY
-    is_correct, errors, warnings = verify_spelling_text(path_input_folder, type_dict_spell)
-    # Numero de erros esperado == 2
-    assert len(errors) == 2
-    # Numero de warnings esperado == 14
-    assert len(warnings) == 14
-
-def test_count_errors_verify_spelling_text_full(): # Teste false
-    path_input_folder = path_input_data_errors
-    type_dict_spell = TypeDict.FULL
-    is_correct, errors, warnings = verify_spelling_text(path_input_folder, type_dict_spell)
-    # Numero de erros esperado == 2
-    assert len(errors) == 2
-    # Numero de warnings esperado == 3
-    assert len(warnings) == 3
-
