+++ tests/unit/test_info.py
@@ -1,0 +1,5 @@
+import src.myparser.info as lv
+
+def test_verstion():
+    ret = lv.print_versions()
+    assert ret is True
