+++ tests/unit/test_structures_files.py
@@ -1,0 +1,18 @@
+from src.orchestrator import verify_structure_folder_files
+
+from tests.unit.test_constants import path_input_data_ground_truth, path_input_data_errors
+
+
+# Testes: Estrutura da pasta de arquivos
+def test_true_verify_structure_folder_files(): # Teste true
+    result_test,__,__ = verify_structure_folder_files(path_input_data_ground_truth)
+    assert result_test is True
+def test_false_verify_structure_folder_files(): # Teste false
+    result_test,errors,warnings = verify_structure_folder_files(path_input_data_errors)
+    assert result_test is False
+def test_count_errors_verify_structure_folder_files(): # Teste false
+    is_correct, errors, warnings = verify_structure_folder_files(path_input_data_errors)
+    # Numero de erros esperado == 2
+    assert len(errors) == 2
+    # Numero de warnings esperado == 0
+    assert len(warnings) == 0
